refactor(funnels): remove non-udf version of trends funnels
