+++ posthog/hogql_queries/insights/funnels/__init__.py
@@ -6,8 +6,6 @@ from .funnel_persons import FunnelActors
 from .funnel_strict import FunnelStrict
 from .funnel_strict_actors import FunnelStrictActors
 from .funnel_time_to_convert import FunnelTimeToConvert
-from .funnel_trends import FunnelTrends
-from .funnel_trends_actors import FunnelTrendsActors
 from .funnel_trends_udf import FunnelTrendsUDF
 from .funnel_udf import FunnelUDF
 from .funnel_unordered import FunnelUnordered
