+++ posthog/hogql_queries/legacy_compatibility/feature_flag.py
@@ -69,27 +69,6 @@ def insight_funnels_use_udf_time_to_convert(team: Team) -> bool:
     )
 
 
-def insight_funnels_use_udf_trends(team: Team) -> bool:
-    return posthoganalytics.feature_enabled(
-        "insight-funnels-use-udf-trends",
-        str(team.uuid),
-        groups={
-            "organization": str(team.organization_id),
-            "project": str(team.id),
-        },
-        group_properties={
-            "organization": {
-                "id": str(team.organization_id),
-            },
-            "project": {
-                "id": str(team.id),
-            },
-        },
-        only_evaluate_locally=False,
-        send_feature_flag_events=False,
-    )
-
-
 def insight_api_use_legacy_queries(team: Team) -> bool:
     """
     Use the legacy implementation of insight api calculation endpoints.
